requirements.txt
Django==4.0.3
djangorestframework==3.13.1
asgiref==3.5.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
mysqlclient==2.1.0
